requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sse-starlette==2.1.3
langgraph==0.2.14
requests==2.32.3
moviepy==1.0.3
langchain-nvidia-ai-endpoints==0.2.2
openai==1.47.1
python-dotenv==1.0.1
pydantic==2.9.2
typing_extensions==4.12.2
numpy==1.26.4
httpx==0.27.2
decorator==4.4.2
